apps/ramadan-bot/requirements.txt
# Ramadan Fajr Bot — Python dependencies
requests>=2.31,<3
pillow>=10.0,<11
streamlit>=1.30,<2
astral>=3.2,<4
pytz>=2024.1
python-dotenv>=1.0,<2
arabic-reshaper>=3.0,<4
python-bidi>=0.6,<1
boto3>=1.34,<2
pytest>=8.2,<9
pytest-cov>=5,<6
playwright>=1.41,<2
